update as per review comments
