+++ src/components/Icons/ParspecLogoIcon.tsx
@@ -1,18 +1,9 @@
-import { SvgIcon, SvgIconProps } from '@mui/material';
-
-function getCustomFontsize(fontSize?: string) {
-    switch (fontSize) {
-        case 'xl':
-            return '40px';
-        case 'xxl':
-            return '44px';
-    }
-}
+import { SvgIconProps } from '@mui/material';
+import { SvgIcon } from './SvgIcon';
 
 export const ParspecLogoIcon: React.FC<SvgIconProps> = ({ fontSize }) => {
-    const customFontSize = getCustomFontsize(fontSize);
     return (
-        <SvgIcon fontSize={fontSize} sx={{ fontSize: customFontSize }} viewBox="0 0 36 36" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <SvgIcon fontSize={fontSize} viewBox="0 0 36 36" fill="none" xmlns="http://www.w3.org/2000/svg">
             <path
                 d="M17.9751 0V5.34674C11.2702 5.36021 5.83898 10.8076 5.83898 17.5254C5.83898 24.2515 11.2837 29.7041 18 29.7041C24.7163 29.7041 30.161 24.2515 30.161 17.5254C30.161 15.2946 29.5621 13.2039 28.5165 11.4057L33.141 8.73189C34.641 11.3166 35.5 14.3205 35.5 17.5254C35.5 27.2044 27.665 35.0508 18 35.0508C8.33502 35.0508 0.5 27.2044 0.5 17.5254C0.5 7.8547 8.32155 0.013468 17.9751 0Z"
                 fill="#FFA72B"
